add Error handling

+++ Application/Activities/Commands/DeleteActivities.cs
@@ -1,4 +1,5 @@
-﻿using Domain;
+﻿using Application.Core;
+using Domain;
 using MediatR;
 using Persistence;
 
@@ -6,12 +7,12 @@ namespace Application.Activities.Commands;
 
 public class DeleteActivities
 {
-    public class Command : IRequest
+    public class Command : IRequest<Result<Unit>>
     {
         public required string Id { get; set; }
     }
 
-    public class Handler : IRequestHandler<Command>
+    public class Handler : IRequestHandler<Command, Result<Unit>>
     {
         private readonly AppDbContext _context;
 
@@ -20,15 +21,22 @@ public class DeleteActivities
             _context = context;
         }
         
-        public async Task Handle(Command request, CancellationToken cancellationToken)
+        public async Task<Result<Unit>> Handle(Command request, CancellationToken cancellationToken)
         {
-            var activity = await _context.Activities
-                .FindAsync([request.Id], cancellationToken)
-                ?? throw new Exception("This activity dont exist");
+            var activity = await _context.Activities.FindAsync([request.Id], cancellationToken);
 
-            _context.Activities.Remove(activity);
-            await _context.SaveChangesAsync(cancellationToken);
+            if (activity == null)
+            {
+                return Result<Unit>.Failure("Activity not found", 404);
+            }
             
+            
+            _context.Activities.Remove(activity);
+            var result = await _context.SaveChangesAsync(cancellationToken)>0;
+
+            if (!result) return Result<Unit>.Failure("Could not delete activity", 400);
+
+            return Result<Unit>.Success(Unit.Value);
         }
     }
 }
